consolidate adam
